+++ lib/node/pl-middle-layer/src/middle_layer/block.ts
@@ -67,7 +67,7 @@ export function blockOutputs(
       const prj = c.accessor(projectEntry).node();
       const ctx = constructBlockContext(projectEntry, blockId);
 
-      return ifNotUndef(getBlockPackInfo(prj, blockId), ({ cfg, bpId }) => {
+      return ifNotUndef(getBlockPackInfo(prj, blockId), ({ cfg, bpId, info }) => {
         const outputs: Record<string, Computable<any>> = {};
         for (const [cellId, cellCfg] of Object.entries(cfg.outputs)) {
           const computableOutput = computableFromCfgOrRF(
@@ -76,6 +76,8 @@ export function blockOutputs(
             cellCfg,
             extractCodeWithInfo(cfg),
             bpId,
+            {},
+            info.source,
           );
           outputs[cellId] = Computable.wrapError(computableOutput, 1);
         }