+++ lib/node/pl-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @milaboratories/pl-client
 
+## 3.16.2
+
+### Patch Changes
+
+- Updated dependencies [e8f26d6]
+  - @milaboratories/pl-model-common@1.49.0
+
+## 3.16.1
+
+### Patch Changes
+
+- 0aa8615: Recognise the `scratchSpace:v1` backend capability, advertised only where the deployment
+  actually has scratch storage.
+
 ## 3.16.0
 
 ### Minor Changes