+++ tools/pl-cli/CHANGELOG.md
@@ -1,5 +1,84 @@
 # @platforma-sdk/pl-cli
 
+## 0.9.15
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.71.8
+
+## 0.9.14
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.71.7
+
+## 0.9.13
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.71.6
+
+## 0.9.12
+
+### Patch Changes
+
+- Updated dependencies [f2ed96c]
+  - @milaboratories/pl-middle-layer@1.71.5
+  - @milaboratories/pl-client@3.16.2
+
+## 0.9.11
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.71.4
+
+## 0.9.10
+
+### Patch Changes
+
+- Updated dependencies [0aa8615]
+  - @milaboratories/pl-client@3.16.1
+  - @milaboratories/pl-middle-layer@1.71.3
+
+## 0.9.9
+
+### Patch Changes
+
+- Updated dependencies [f532ce7]
+  - @milaboratories/pl-middle-layer@1.71.2
+
+## 0.9.8
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.71.1
+
+## 0.9.7
+
+### Patch Changes
+
+- Updated dependencies [9a6d54d]
+  - @milaboratories/pl-middle-layer@1.71.0
+
+## 0.9.6
+
+### Patch Changes
+
+- Updated dependencies [d8b9666]
+  - @milaboratories/pl-middle-layer@1.70.0
+
+## 0.9.5
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.69.2
+
+## 0.9.4
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.69.1
+
 ## 0.9.3
 
 ### Patch Changes